Implemented VR_MainController + VRTheWorld
Added new versions for VR capability so that we can change where necessary dependent on the VR. So far everything now works as if on the desktop. Next steps would be to add the teleportation, locomotion/movement of user in the game (immersive effect). Then adding the controller mobility or how to enable "touch" capability for the buttons and joystick. Also need to add more objects to pick up.